fix(payment): BOLT-77 Fix the issue with store styling
